@@ -6,10 +6,9 @@ use RuntimeException;
 use SlmQueueDoctrine\Exception\ExceptionInterface;
 
 /**
- * ReleasableException. Throw this exception in the "execute" method of your job so that the worker
- * puts back the job into the queue
+ * BuryableException
  */
-class ReleasableException extends RuntimeException implements ExceptionInterface
+class BuryableException extends RuntimeException implements ExceptionInterface
 {
     /**
      * @var array
@@ -18,12 +17,12 @@ class ReleasableException extends RuntimeException implements ExceptionInterface
 
     /**
      * Valid options are:
-     *      - scheduled: the time when the job should run the next time
-     *      - delay: the delay in seconds before a job become available to be popped (default to 0 - no delay -)
+     *      - message: Message why this has happened
+     *      - trace: Stack trace for further investigation
      *
      * @param array $options
      */
-    public function __construct(array $options = array())
+    public function __construct(array $options = [])
     {
         $this->options = $options;
     }
