@@ -3,8 +3,10 @@
 namespace SlmQueueDoctrineTest\Controller;
 
 use PHPUnit_Framework_TestCase as TestCase;
+use SlmQueueDoctrine\Controller\DoctrineWorkerController;
 use SlmQueueDoctrineTest\Util\ServiceManagerFactory;
 use Zend\Mvc\Router\RouteMatch;
+use Zend\ServiceManager\Exception\ServiceNotFoundException;
 use Zend\ServiceManager\ServiceManager;
 
 class DoctrineWorkerControllerTest extends TestCase
@@ -22,11 +24,11 @@ class DoctrineWorkerControllerTest extends TestCase
 
     public function testThrowExceptionIfQueueIsUnknown()
     {
-        $controller = $this->serviceManager->get('ControllerLoader')->get('SlmQueueDoctrine\Controller\DoctrineWorkerController');
-        $routeMatch = new RouteMatch(array('queue' => 'unknownQueue'));
+        $controller = $this->serviceManager->get('ControllerLoader')->get(DoctrineWorkerController::class);
+        $routeMatch = new RouteMatch(['queue' => 'unknownQueue']);
         $controller->getEvent()->setRouteMatch($routeMatch);
 
-        $this->setExpectedException('Zend\ServiceManager\Exception\ServiceNotFoundException');
+        $this->setExpectedException(ServiceNotFoundException::class);
         $controller->processAction();
     }
 }
