@@ -3,7 +3,9 @@
 namespace SlmQueueDoctrineTest\Factory;
 
 use PHPUnit_Framework_TestCase;
+use SlmQueue\Queue\QueuePluginManager;
 use SlmQueueDoctrine\Factory\DoctrineQueueFactory;
+use SlmQueueDoctrine\Queue\DoctrineQueue;
 use SlmQueueDoctrineTest\Util\ServiceManagerFactory;
 
 class DoctrineQueueFactoryTest extends PHPUnit_Framework_TestCase
@@ -11,17 +13,17 @@ class DoctrineQueueFactoryTest extends PHPUnit_Framework_TestCase
     public function testCreateServiceGetsInstance()
     {
         $sm                 = ServiceManagerFactory::getServiceManager();
-        $queuePluginManager = $sm->get('SlmQueue\Queue\QueuePluginManager');
+        $queuePluginManager = $sm->get(QueuePluginManager::class);
         $factory            = new DoctrineQueueFactory();
         $service            = $factory->createService($queuePluginManager);
 
-        $this->assertInstanceOf('SlmQueueDoctrine\Queue\DoctrineQueue', $service);
+        $this->assertInstanceOf(DoctrineQueue::class, $service);
     }
 
     public function testSpecifiedQueueOptionsOverrideModuleDefaults()
     {
         $sm                 = ServiceManagerFactory::getServiceManager();
-        $queuePluginManager = $sm->get('SlmQueue\Queue\QueuePluginManager');
+        $queuePluginManager = $sm->get(QueuePluginManager::class);
         $config             = $sm->get('config');
 
         $factory            = new DoctrineQueueFactory();
