@@ -19,6 +19,7 @@
 
 namespace SlmQueueDoctrineTest\Util;
 
+use Zend\Mvc\Service\ServiceListenerFactory;
 use Zend\ServiceManager\ServiceManager;
 use Zend\Mvc\Service\ServiceManagerConfig;
 
@@ -50,11 +51,10 @@ class ServiceManagerFactory
     public static function getServiceManager()
     {
         $serviceManager = new ServiceManager(new ServiceManagerConfig(
-            isset(static::$config['service_manager']) ? static::$config['service_manager'] : array()
+            isset(static::$config['service_manager']) ? static::$config['service_manager'] : []
         ));
         $serviceManager->setService('ApplicationConfig', static::$config);
-        $serviceManager->setFactory('ServiceListener', 'Zend\Mvc\Service\ServiceListenerFactory');
-        $serviceManager->setFactory('ModuleListener', 'Zend\Mvc\Service\ModuleListenerFactory');
+        $serviceManager->setFactory('ServiceListener', ServiceListenerFactory::class);
 
         /** @var $moduleManager \Zend\ModuleManager\ModuleManager */
         $moduleManager = $serviceManager->get('ModuleManager');
